frogfs_get_name fix
frogfs_get_name was orignally an internal function that got exposed as an api
function.  NULL terminating its string when the size is a multiple of 4 was
overlooked, so a small API change was required to ensure it always is NULL
terminated.

+++ src/frogfs_priv.h
@@ -36,7 +36,7 @@ typedef struct frogfs_fh_t {
 typedef struct frogfs_dh_t {
     const frogfs_fs_t *fs; /**< frogfs fs pointer */
     const frogfs_dir_t *dir; /**< frogfs entry */
-    uint16_t index; /**< current index */
+    long index; /**< current index */
 } frogfs_dh_t;
 
 /**
